produce shared library instead of module (#19)
E.g. https://github.com/KhronosGroup/OpenCL-Layers/blob/main/object-lifetime/CMakeLists.txt also uses that. It doesn't matter much on Linux and Windows, but on MacOS currently a `.so` file is produced rather than `.dylib` which is a bit unusual and a bit of a pain to package. I've tested that both versions work though, would you be ok with changing this?